requirements.txt
google-adk>=1.0.0
google-cloud-aiplatform[agent_engines,adk]>=1.112.0
google-cloud-speech>=2.20.0
google-cloud-storage>=2.10.0
streamlit>=1.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sendgrid>=6.10.0